Expose parser_edit_set_variable()
Also make it only set the first occurrence of a variable.  As a
side effect this also prevents it from pasting the new value multiple
times after every VARIABLE_TOKEN of the variable.

This is incomplete.  The value needs to go through the parser too
and not just injected as is or things will be badly formatted in the end.

Signed-off-by: Tobias Kortkamp <tobik@FreeBSD.org>