Lookup defined port options
This is used for determining what is an options helper and what is
not.  This opens a whole new class of checks for portclippy/portscan.

Since this might get in the way when creating new ports when options
have not been added to OPTIONS_DEFINE yet, portfmt(1) and portedit
merge retain the old behavior of dynamically determining this based
on if the helper matches a particular pattern.  This is gated behind
the new PARSER_DYNAMIC_PORT_OPTIONS behavior setting.

Signed-off-by: Tobias Kortkamp <tobik@FreeBSD.org>